finance/plaid: move + split plaid_utils into db / sync / link
plaid_utils joins //finance, split by role (it's not MCP -- the Postgres MCP in
front of the mirror is the generic agents/plaid-db-mcp):
- //finance/plaid/db   -- mirror library: schema, migrations, client, read_model,
  sync logic, link_store/profiles, secret_store + shared PlaidWebSettings (config.py).
- //finance/plaid/sync -- cron full-refresh sync (sync_cli + run_sync) -> plaid-mcp-sync image.
- //finance/plaid/link -- FastAPI Link-UI web app (create_app + routes) -> plaid-mcp-server image.

git mv mcp_server->link, plaid_utils->db, sync_cli->sync; extracted PlaidWebSettings/
MAX_TRANSACTION_DAYS->db/config.py and run_sync->sync/sync_cli.py; codemod
plaid_utils.*->finance.plaid.db.* + //plaid_utils->//finance/plaid/db; push-images
image targets -> //finance/plaid/{link,sync} (ghcr names unchanged); widened
db visibility to //finance/plaid:__subpackages__.

Validated locally: build clean; tests pass (db client/link_profiles/link_store/
read_model/sync, link test_app, augur server_test + budget test_sql_read_model).
