+++ finance/beancount_export/runner.py
@@ -29,7 +29,7 @@ from beancount.parser import printer
 from finance.augur.budget.schema import BudgetConfig
 from finance.augur.budget.sql_read_model import ClassifiedRow, read_all_classified
 from finance.beancount_export.export import ClassifiedTxn, render_ledger
-from plaid_utils.schema import async_session_factory
+from finance.plaid.db.schema import async_session_factory
 
 log = structlog.get_logger()
