+++ finance/plaid/db/BUILD.bazel
@@ -3,13 +3,13 @@ load("//devinfra/python:defs.bzl", "py_binary", "py_library", "py_test")
 py_library(
     name = "link_profiles",
     srcs = ["link_profiles.py"],
-    visibility = ["//plaid_utils:__subpackages__"],
+    visibility = ["//finance/plaid:__subpackages__"],
 )
 
 py_library(
     name = "client",
     srcs = ["client.py"],
-    visibility = ["//plaid_utils:__subpackages__"],
+    visibility = ["//finance/plaid:__subpackages__"],
     deps = [
         ":link_profiles",
         "@pypi//certifi",
@@ -45,7 +45,7 @@ py_library(
     visibility = [
         "//finance/augur:__subpackages__",
         "//finance/beancount_export:__pkg__",
-        "//plaid_utils:__subpackages__",
+        "//finance/plaid:__subpackages__",
     ],
     deps = [
         "@pypi//asyncpg",
@@ -141,7 +141,7 @@ py_library(
     name = "link_store",
     srcs = ["link_store.py"],
     data = glob(["migrations/**"]),
-    visibility = ["//plaid_utils:__subpackages__"],
+    visibility = ["//finance/plaid:__subpackages__"],
     deps = [
         ":link_profiles",
         ":schema",
@@ -154,19 +154,30 @@ py_library(
 py_library(
     name = "secret_store",
     srcs = ["secret_store.py"],
-    visibility = ["//plaid_utils:__subpackages__"],
+    visibility = ["//finance/plaid:__subpackages__"],
     deps = ["@pypi//kubernetes_asyncio"],
 )
 
 py_library(
     name = "sync",
     srcs = ["sync.py"],
-    visibility = ["//plaid_utils:__subpackages__"],
+    visibility = ["//finance/plaid:__subpackages__"],
     deps = [
         ":link_profiles",
         ":link_store",
         ":secret_store",
         "@pypi//plaid_python",
+    ],
+)
+
+py_library(
+    name = "config",
+    srcs = ["config.py"],
+    visibility = ["//finance/plaid:__subpackages__"],
+    deps = [
+        ":sync",
+        "@pypi//pydantic",
+        "@pypi//pydantic_settings",
     ],
 )
 
@@ -182,6 +193,6 @@ py_library(
 
 py_binary(
     name = "sandbox_smoke",
-    main_module = "plaid_utils.sandbox_smoke",
+    main_module = "finance.plaid.db.sandbox_smoke",
     deps = [":sandbox_smoke_lib"],
 )
