+++ finance/plaid/link/BUILD.bazel
@@ -3,51 +3,36 @@ load("@rules_oci//oci:defs.bzl", "oci_image", "oci_load")
 load("//devinfra/python:defs.bzl", "py_binary", "py_library", "py_test")
 load("//props:oci.bzl", "py_image_entrypoint", "py_image_env")
 
+# Plaid Link-UI web app (FastAPI). Not an MCP server; serves the account-linking
+# UI + link-management API. The cron sync process is //finance/plaid/sync.
+
 py_library(
     name = "app",
     srcs = ["app.py"],
     deps = [
-        "//plaid_utils:client",
-        "//plaid_utils:link_profiles",
-        "//plaid_utils:link_store",
-        "//plaid_utils:secret_store",
-        "//plaid_utils:sync",
+        "//finance/plaid/db:client",
+        "//finance/plaid/db:config",
+        "//finance/plaid/db:link_profiles",
+        "//finance/plaid/db:link_store",
+        "//finance/plaid/db:secret_store",
+        "//finance/plaid/db:sync",
         "@pypi//fastapi",
         "@pypi//plaid_python",
         "@pypi//pydantic",
-        "@pypi//pydantic_settings",
         "@pypi//uvicorn",
     ],
 )
 
 py_binary(
     name = "app_cli",
-    main_module = "plaid_utils.mcp_server.app",
+    main_module = "finance.plaid.link.app",
     deps = [":app"],
-)
-
-py_library(
-    name = "sync_cli_lib",
-    srcs = ["sync_cli.py"],
-    deps = [":app"],
-)
-
-py_binary(
-    name = "sync_cli",
-    main_module = "plaid_utils.mcp_server.sync_cli",
-    deps = [":sync_cli_lib"],
 )
 
 aspect_py_binary(
     name = "server_bin",
     main = "app.py",
     deps = [":app"],
-)
-
-aspect_py_binary(
-    name = "sync_bin",
-    main = "sync_cli.py",
-    deps = [":sync_cli_lib"],
 )
 
 py_image_layer(
@@ -74,38 +59,15 @@ oci_load(
     repo_tags = ["plaid-mcp-server:latest"],
 )
 
-py_image_layer(
-    name = "sync_layers",
-    binary = ":sync_bin",
-    group = "1000",
-    owner = "1000",
-)
-
-oci_image(
-    name = "sync_image",
-    base = "@debian_slim_linux_amd64",
-    entrypoint = py_image_entrypoint("sync_bin"),
-    env = py_image_env(binary_name = "sync_bin"),
-    labels = {"org.opencontainers.image.source": "https://github.com/agentydragon/ducktape"},
-    tars = [":sync_layers"],
-    user = "1000:1000",
-    visibility = ["//visibility:public"],
-)
-
-oci_load(
-    name = "sync_load",
-    image = ":sync_image",
-    repo_tags = ["plaid-mcp-sync:latest"],
-)
-
 py_test(
     name = "test_app",
     srcs = ["test_app.py"],
     deps = [
         ":app",
-        "//plaid_utils:client",
-        "//plaid_utils:link_profiles",
-        "//plaid_utils:link_store",
+        "//finance/plaid/db:client",
+        "//finance/plaid/db:config",
+        "//finance/plaid/db:link_profiles",
+        "//finance/plaid/db:link_store",
         "@pypi//fastapi",
         "@pypi//httpx",
         "@pypi//plaid_python",
